+++ Resources/Locale/ru-RU/ss220/Command/command.ftl
@@ -2,3 +2,5 @@ command-makerandomantag-description = Делает случайного игро
 command-makerandomantag-negative = Не повезло.
 command-makerandomantag-objective = Необходимо указать тип антагониста из предложенных.
 command-makerandomantag-sucess = {$Entityname} успешно стал {$antag}
+
+command-garbage-clear-up-desc = Удаляет весь мусор с карты (применимо к объектам с тегом 'trash')
